Move remaining edit routes to edit folder
