+++ client/src/components/edit/UploadGallery.js
@@ -1,6 +1,6 @@
 import React, {useState, useContext} from 'react'
-import AppContext from "../context/AppContext"
-import AlertContext from "../context/alertContext"
+import AppContext from "../../context/AppContext"
+import AlertContext from "../../context/alertContext"
 import axios from 'axios'
 import { CSSTransition, TransitionGroup} from 'react-transition-group';
