+++ client/src/pages/Edit.js
@@ -7,10 +7,11 @@ import "../styles/edit.css"
 import AppContext from "../context/AppContext"
 import AuthContext from "../context/authContext"
 
-import UploadGallery from '../components/UploadGallery'
-import UpdateStock from '../components/UpdateStock'
-import UploadPrint from "../components/UploadPrint"
+import UploadGallery from '../components/edit/UploadGallery'
+import UpdateStock from '../components/edit/UpdateStock'
+import UploadPrint from "../components/edit/UploadPrint"
 import EditGallery from "../components/edit/EditGallery"
+
 import PageHeader from '../components/layout/PageHeader'
 import Alerts from "../components/layout/Alerts"
